refactory at deploy scripts
