+++ bridge/deploy/02_deploy_proxyAdmin.js
@@ -2,20 +2,19 @@ module.exports = async function({getNamedAccounts, deployments}) { // HardhatRun
   const {deployer, multiSig, proxyAdmin} = await getNamedAccounts();
   const {deploy, log, execute} = deployments;
 
-  if (proxyAdmin)  return
+  if (proxyAdmin) return;
+
   const deployResult = await deploy('ProxyAdmin', {
     from: deployer,
     log: true
   });
 
-  if (deployResult.newlyDeployed) {
-    log(
-      `Contract ProxyAdmin deployed at ${deployResult.address} using ${deployResult.receipt.gasUsed.toString()} gas`
-    );
-    const MultiSigWallet = await deployments.get('MultiSigWallet');
-    await execute('ProxyAdmin', {from: deployer}, 'transferOwnership', multiSig ?? MultiSigWallet.address);
-    log(`Transfered Ownership to MultiSig`);
-  }
+  if (!deployResult.newlyDeployed) return;
+
+  log(`Contract ProxyAdmin deployed at ${deployResult.address} using ${deployResult.receipt.gasUsed.toString()} gas`);
+  const MultiSigWallet = await deployments.get('MultiSigWallet');
+  await execute('ProxyAdmin', {from: deployer}, 'transferOwnership', multiSig ?? MultiSigWallet.address);
+  log(`Transfered Ownership to MultiSig`);
 };
 module.exports.id = 'deploy_proxyAdmin'; // id required to prevent reexecution
 module.exports.tags = ['ProxyAdmin'];
