+++ bridge/deploy/03_deploy_sideTokenFactory.js
@@ -1,20 +1,20 @@
 module.exports = async function ({getNamedAccounts, deployments}) { // HardhatRuntimeEnvironment
-    const {deployer} = await getNamedAccounts()
-    const {deploy, log} = deployments
+  const {deployer, bridgeProxy} = await getNamedAccounts();
+  const {deploy, log} = deployments;
 
-    if (sideTokenFactory) return
-    const deployResult = await deploy('SideTokenFactory', {
-      from: deployer,
-      log: true,
-    });
+  // if we have the bridge proxy address
+  // doesn't need to deploy a new SideTokenFactory
+  // because most probably the sideTokenFactory already exists
+  if (bridgeProxy) return;
 
-    if (deployResult.newlyDeployed) {
-      log(
-        `Contract SideTokenFactory deployed at ${deployResult.address} using ${deployResult.receipt.gasUsed.toString()} gas`
-      );
-    }
+  const deployResult = await deploy('SideTokenFactory', {
+    from: deployer,
+    log: true,
+  });
 
+  if (deployResult.newlyDeployed) {
+    log(`Contract SideTokenFactory deployed at ${deployResult.address} using ${deployResult.receipt.gasUsed.toString()} gas`);
+  }
 };
 module.exports.id = 'deploy_sideTokenFactory'; // id required to prevent reexecution
 module.exports.tags = ['SideTokenFactory', 'new'];
-module.exports.dependencies = ['BridgeProxy'];
