--- bridge/deploy/12_conditional_deploy_mainToken.js
@@ -1,24 +1,0 @@
-module.exports = async function ({getNamedAccounts, deployments, network}) { // HardhatRuntimeEnvironment
-    const {deployer} = await getNamedAccounts()
-    const {deploy, log} = deployments
-
-    if (!network.live) {
-        const deployResult = await deploy('MainToken', {
-            from: deployer,
-            args: [
-                'MAIN',
-                'MAIN',
-                18,
-                web3.utils.toWei('1000')
-            ],
-            log: true,
-        });
-        if (deployResult.newlyDeployed) {
-            log(
-                `Contract MainToken deployed at ${deployResult.address} using ${deployResult.receipt.gasUsed.toString()} gas`
-            );
-        }
-    }
-};
-module.exports.id = 'deploy_mainToken'; // id required to prevent reexecution
-module.exports.tags = ['MainToken', 'test'];
