+++ bridge/deployments/development/.chainId
@@ -1,0 +1,1 @@
+1337
